feat(reflection): serialize reflected byte blobs atomically
Add explicit std::byte and Blob reflection, preserve bounded transactional archive and package behavior, and migrate volume texture source voxels with versioned Array compatibility. Qualify the real 128-cubed horizontal PNG import, save, reload, reimport, and cook workflow.

Plan: Documentation/Plans/ReflectedByteBlobSerialization.md

Stage: Stage 4: Qualify and publish the lasting contract

+++ Engine/Tests/Native/EngineTests/Private/VolumetricCloudSceneVulkanTests.cpp
@@ -112,7 +112,7 @@ namespace Durin
 		{
 			auto* Texture = NewObject<DVolumeTexture>(nullptr, Name);
 			FVolumeTextureSourceData Source{
-				.Voxels = {Density}, .Width = 1, .Height = 1, .Depth = 1,
+				.Voxels = {static_cast<std::byte>(Density)}, .Width = 1, .Height = 1, .Depth = 1,
 				.Format = EVolumeTextureFormat::R8_UNORM};
 			auto Platform = std::make_unique<FVolumeTexturePlatformData>();
 			Platform->PixelFormat = EPixelFormat::R8_UNORM;
